+++ cxflow_tf/__init__.py
@@ -1,2 +1,3 @@
 from .tensorboard_hook import TensorBoardHook
-from .tf_net import BaseTFNet, BaseTFNetRestore, create_activation, create_optimizer
+from .tf_net import BaseTFNet, BaseTFNetRestore
+from .utils import TF_OPTIMIZERS_MODULE, create_activation, create_optimizer, repeat
